fix delete project users

+++ backend/app/Service/CustomerProject/CustomerProjectService.php
@@ -3,6 +3,7 @@
 namespace App\Service\CustomerProject;
 
 use App\Models\CustomerProject;
+use App\Models\Ticket;
 
 class CustomerProjectService implements CustomerProjectServiceInterface
 {
@@ -20,7 +21,17 @@ class CustomerProjectService implements CustomerProjectServiceInterface
 
     public function delete($id)
     {
-        $data = CustomerProject::where('id', $id)->first();
+        // $data = CustomerProject::where('id', $id)->first();
+
+        // return $data->delete();
+
+        $data = CustomerProject::find($id);
+
+        $userIsInTicket = Ticket::where('customer_project_id', $id)->exists();
+
+        if( $userIsInTicket ) {
+            return false;
+        }
 
         return $data->delete();
     }
